refactor: rename install-banner.tsx to InstallBanner.tsx
Follow PascalCase convention for React components per CLAUDE.md.
