+++ resources/js/components/shared/InstallBanner.tsx
